Update deploy.sh
The "cd: OLDPWD not set" warning occurs when the script tries to return to the previous directory using cd - but there's no previous directory stored in the OLDPWD variable. This is a minor issue and can be fixed by simply saving the original directory before changing to the temporary directory.